Fix squashing of pear exit codes
Commit 286125b fixed an error from where badly encoded data would cause crashes,
but the fix of piping it through iconv mean that the status code returned was
that of iconv. This causes the module to always treat the pear module as
"installed," meaning the install exec would never run. This changes fixes that
by sending the standard output to /dev/null, which preserves the original exit
code.